// src/app/features/suppliers/models/supplier.ts
export interface Supplier {

    id: number;

    supplierCode: string;

    companyName: string;

    contactPerson: string;

    email: string;

    phone: string;

    address: string;

    city: string;

    state: string;

    country: string;

    postalCode: string;

    isActive: boolean;

}